Isolate registration component, add admin password check for late regi
