+++ .env.template
@@ -3,6 +3,7 @@ NEXT_PUBLIC_SUPABASE_ANON_KEY=
 SUPABASE_SERVER_KEY=
 NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=
 STRIPE_SECRET_KEY=
+NEXT_PUBLIC_LATE_REGISTRATION=
 
 ========= Price IDs =========
